+++ inlang/source-code/paraglide/paraglide-sveltekit/src/vite/preprocessor/types.d.ts
@@ -12,7 +12,7 @@ export type ElementNode<Name extends stirng> = {
 	? {
 			name: "svelte:element"
 			tag: string | Expression
-	  }
+		}
 	: { name: Name })
 
 type Expression = {