+++ inlang/source-code/website/src/pages/index/custom_section/HeroSearch.tsx
@@ -76,8 +76,8 @@ function Play() {
 
 function addLinksToText(text: string) {
 	const replacements: { [key: string]: string } = {
+		"lix apps": `https://github.com/opral/monorepo/tree/main/lix`,
 		"use case": `${isProduction ? `https://inlang.com?` : "http://localhost:3000?"}#personas`,
-		"change control": `${isProduction ? `https://inlang.com?` : "http://localhost:3000?"}#lix`,
 	}
 
 	const elements = []