+++ inlang/source-code/paraglide/paraglide-vite/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @inlang/paraglide-vite
 
+## 1.2.57
+
+### Patch Changes
+
+- @inlang/paraglide-unplugin@1.6.7
+
+## 1.2.56
+
+### Patch Changes
+
+- @inlang/paraglide-unplugin@1.6.6
+
 ## 1.2.55
 
 ### Patch Changes