+++ inlang/source-code/plugins/json/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @inlang/plugin-json
 
+## 5.1.41
+
+### Patch Changes
+
+- Updated dependencies [d27a983]
+- Updated dependencies [a27b7a4]
+  - @inlang/sdk@0.34.10
+
 ## 5.1.40
 
 ### Patch Changes