+++ inlang/source-code/plugins/t-function-matcher/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @inlang/plugin-t-function-matcher
 
+## 0.6.20
+
+### Patch Changes
+
+- Updated dependencies [d27a983]
+- Updated dependencies [a27b7a4]
+  - @inlang/sdk@0.34.10
+
 ## 0.6.19
 
 ### Patch Changes